separate reconsideration into module
