+++ alcs-api/src/application-reconsideration/application-reconsideration.entity.ts
@@ -1,9 +1,9 @@
 import { AutoMap } from '@automapper/classes';
 import { Type } from 'class-transformer';
 import { Column, Entity, JoinColumn, ManyToOne, OneToOne } from 'typeorm';
-import { Card } from '../../card/card.entity';
-import { Base } from '../../common/entities/base.entity';
-import { Application } from '../application.entity';
+import { Application } from '../application/application.entity';
+import { Card } from '../card/card.entity';
+import { Base } from '../common/entities/base.entity';
 import { ApplicationReconsiderationType } from './reconsideration-type/application-reconsideration-type.entity';
 
 @Entity()
