+++ alcs-api/src/application-reconsideration/applicationReconsideration.dto.ts
@@ -6,11 +6,11 @@ import {
   IsOptional,
   IsString,
 } from 'class-validator';
-import { CardDto } from '../../card/card.dto';
-import { ApplicationRegionDto } from '../../code/application-code/application-region/application-region.dto';
-import { ApplicationTypeDto } from '../../code/application-code/application-type/application-type.dto';
-import { BaseCodeDto } from '../../common/dtos/base.dto';
-import { ApplicationDecisionMeetingDto } from '../application-decision-meeting/application-decision-meeting.dto';
+import { ApplicationDecisionMeetingDto } from '../application/application-decision-meeting/application-decision-meeting.dto';
+import { CardDto } from '../card/card.dto';
+import { ApplicationRegionDto } from '../code/application-code/application-region/application-region.dto';
+import { ApplicationTypeDto } from '../code/application-code/application-type/application-type.dto';
+import { BaseCodeDto } from '../common/dtos/base.dto';
 
 export class ReconsiderationTypeDto extends BaseCodeDto {}
