+++ alcs-api/src/application-reconsideration/reconsideration-type/application-reconsideration-type.entity.ts
@@ -1,5 +1,5 @@
 import { Entity } from 'typeorm';
-import { BaseCodeEntity } from '../../../common/entities/base.code.entity';
+import { BaseCodeEntity } from '../../common/entities/base.code.entity';
 
 @Entity()
 export class ApplicationReconsiderationType extends BaseCodeEntity {}
