+++ alcs-api/src/card/card-subtask/card-subtask.dto.ts
@@ -1,6 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { IsOptional, IsUUID } from 'class-validator';
-import { ApplicationReconsiderationDto } from '../../application/application-reconsideration/applicationReconsideration.dto';
+import { ApplicationReconsiderationDto } from '../../application-reconsideration/applicationReconsideration.dto';
 import { ApplicationDto } from '../../application/application.dto';
 
 export class UpdateCardSubtaskDto {
