+++ alcs-api/src/code/application-code/application-code.dto.ts
@@ -1,4 +1,4 @@
-import { ReconsiderationTypeDto } from '../../application/application-reconsideration/applicationReconsideration.dto';
+import { ReconsiderationTypeDto } from '../../application-reconsideration/applicationReconsideration.dto';
 import { CardStatusDto } from '../../card/card-status/card-status.dto';
 import { ApplicationRegionDto } from './application-region/application-region.dto';
 import { ApplicationTypeDto } from './application-type/application-type.dto';
