+++ alcs-api/src/code/code.controller.ts
@@ -4,8 +4,8 @@ import { Controller, Get, UseGuards } from '@nestjs/common';
 import { ApiOAuth2 } from '@nestjs/swagger';
 import * as config from 'config';
 import { RoleGuard } from 'nest-keycloak-connect';
-import { ReconsiderationTypeDto } from '../application/application-reconsideration/applicationReconsideration.dto';
-import { ApplicationReconsiderationType } from '../application/application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
+import { ReconsiderationTypeDto } from '../application-reconsideration/applicationReconsideration.dto';
+import { ApplicationReconsiderationType } from '../application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
 import { CardStatusDto } from '../card/card-status/card-status.dto';
 import { CardStatus } from '../card/card-status/card-status.entity';
 import { ANY_AUTH_ROLE } from '../common/authorization/roles';
