+++ alcs-api/src/common/automapper/reconsideration.automapper.profile.ts
@@ -1,17 +1,17 @@
 import { createMap, forMember, mapFrom, Mapper } from '@automapper/core';
 import { AutomapperProfile, InjectMapper } from '@automapper/nestjs';
 import { Injectable } from '@nestjs/common';
-import { ApplicationDecisionMeetingDto } from '../../application/application-decision-meeting/application-decision-meeting.dto';
-import { ApplicationDecisionMeeting } from '../../application/application-decision-meeting/application-decision-meeting.entity';
-import { ApplicationReconsideration } from '../../application/application-reconsideration/application-reconsideration.entity';
+import { ApplicationReconsideration } from '../../application-reconsideration/application-reconsideration.entity';
 import {
   ApplicationForReconsiderationDto,
   ApplicationReconsiderationCreateDto,
   ApplicationReconsiderationDto,
   ApplicationReconsiderationWithoutApplicationDto,
   ReconsiderationTypeDto,
-} from '../../application/application-reconsideration/applicationReconsideration.dto';
-import { ApplicationReconsiderationType } from '../../application/application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
+} from '../../application-reconsideration/applicationReconsideration.dto';
+import { ApplicationReconsiderationType } from '../../application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
+import { ApplicationDecisionMeetingDto } from '../../application/application-decision-meeting/application-decision-meeting.dto';
+import { ApplicationDecisionMeeting } from '../../application/application-decision-meeting/application-decision-meeting.entity';
 import { Application } from '../../application/application.entity';
 
 @Injectable()
