+++ alcs-api/src/common/utils/test-helpers/mockEntities.ts
@@ -1,10 +1,10 @@
+import { ApplicationReconsideration } from '../../../application-reconsideration/application-reconsideration.entity';
+import { ApplicationReconsiderationType } from '../../../application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
 import { ApplicationDecisionMeeting } from '../../../application/application-decision-meeting/application-decision-meeting.entity';
 import { ApplicationDecisionOutcomeType } from '../../../application/application-decision/application-decision-outcome.entity';
 import { ApplicationDecision } from '../../../application/application-decision/application-decision.entity';
 import { ApplicationMeeting } from '../../../application/application-meeting/application-meeting.entity';
 import { ApplicationPaused } from '../../../application/application-paused.entity';
-import { ApplicationReconsideration } from '../../../application/application-reconsideration/application-reconsideration.entity';
-import { ApplicationReconsiderationType } from '../../../application/application-reconsideration/reconsideration-type/application-reconsideration-type.entity';
 import { Application } from '../../../application/application.entity';
 import { Board } from '../../../board/board.entity';
 import { CardStatus } from '../../../card/card-status/card-status.entity';
